requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
PyGithub==2.1.1
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
google-auth==2.23.4
gunicorn==21.2.0
requests==2.31.0